% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Perform data checks on IDDO-SDTM data}
\usage{
check_data(data, age_in_years = FALSE)
}
\arguments{
\item{data}{A dataset using IDDO-SDTM columns}

\item{age_in_years}{Boolean. Is the AGE column in years. Default is FALSE,
and function will call \code{convert_age_to_years} if FALSE, otherwise will not
convert age.}
}
\value{
A list with various summaries/checks, the number of which is
dependent on what variables are present in the input data. A plot is also
returned showing the missingness in each variable.
}
\description{
Provides a variety of checks and summaries on IDDO-SDTM data including
summarising the number of rows by study ID (STUDYID), the number
of participants under 6 month, 18 years and over 90 years, and the number of
units used by each test (TESTCD) and study ID.
}
\examples{
check_data(DM_RPTESTB)

check_data(LB_RPTESTB)

}
