% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clinical_table.R
\name{create_clinical_table}
\alias{create_clinical_table}
\title{Creates table of key clinical features}
\usage{
create_clinical_table(
  dm_domain,
  mb_domain = NULL,
  mp_domain = NULL,
  sa_domain = NULL,
  vs_domain = NULL
)
}
\arguments{
\item{dm_domain}{The name of the demographics (DM) domain in the global
environment.}

\item{mb_domain}{The name of the microbiology (MB) domain in the global
environment.}

\item{mp_domain}{The name of the morphology and physiology (MP) domain in the
global environment.}

\item{sa_domain}{The name of the clinical and adverse events (SA) domain in
the global environment.}

\item{vs_domain}{The name of the vital signs (VS) domain in the global
environment.}
}
\value{
An analysis dataset, one row per person, per timepoint.
}
\description{
Uses several IDDO-SDTM domains to create a single dataset with various
demographic, microbiology, physiology, clinical and vital sign information
summarised.
}
