% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_malaria_pcr_table.R
\name{create_malaria_pcr_table}
\alias{create_malaria_pcr_table}
\title{Create table of malaria specific PCR data}
\usage{
create_malaria_pcr_table(pf_domain, rs_domain, values_funct = first)
}
\arguments{
\item{pf_domain}{A pharmacogenomics genetics/PF domain data frame}

\item{rs_domain}{A disease response and clinical classification/RS domain
data frame.}

\item{values_funct}{Function. The function which will determine which data row is used
in the output, in the event there are multiple rows for the same subject
with the same time points (as listed in timing_variables). Default is
first(), i.e. if there is two rows from the same day and time, the first
record will be taken, the second will be dropped. Choice of
timing_variables will impact the number of rows affected.}
}
\value{
An analysis dataset, one row per person, per timepoint.
}
\description{
Joins the pharmacogenomics genetics (PF) and disease response and clinical
classification (RS) IDDO-SDTM domains together to create a single dataset.
}
\examples{
create_malaria_pcr_table(PF_RPTESTB, RS_RPTESTB)

}
