% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_participant_table.R
\name{create_participant_table}
\alias{create_participant_table}
\title{Create table of participant details and baseline characteristics.}
\usage{
create_participant_table(
  dm_domain,
  lb_domain = NULL,
  mb_domain = NULL,
  rp_domain = NULL,
  sc_domain = NULL,
  vs_domain = NULL,
  print_messages = TRUE
)
}
\arguments{
\item{dm_domain}{A demographics/DM domain data frame.}

\item{lb_domain}{A laboratory/LB domain data frame.}

\item{mb_domain}{A microbiology/MB domain data frame.}

\item{rp_domain}{A reproductive system findings/RP domain data frame.}

\item{sc_domain}{A subject characteristics/SC domain data frame.}

\item{vs_domain}{A vital signs/VS domain data frame.}

\item{print_messages}{Boolean. Should messages from the function be generated
and shown in the user's console. Default is TRUE.}
}
\value{
An analysis dataset, one row per participant.
}
\description{
Joins several IDDO-SDTM domains together to create a single dataset with
participant demographics, characteristics and baseline test results or
findings. Baseline timing is defined as planned
study day (VISITDY) = 1 or epoch (EPOCH) = BASELINE. If height, weight, age
and sex are available, then anthro growth standards will be computed
for under 5 year olds
}
\examples{
create_participant_table(dm_domain = DM_RPTESTB,
                         lb_domain = LB_RPTESTB,
                         vs_domain = VS_RPTESTB)


}
