% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HO_RPTESTB.R
\docType{data}
\name{HO_RPTESTB}
\alias{HO_RPTESTB}
\title{IDDO-SDTM Tutorial Healthcare Encounters (HO) Domain}
\format{
\subsection{}{

A data frame with 3 rows and 24 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{HOSEQ}{Sequence Number}
\item{HOTERM}{Reported Term for the Disposition Event}
\item{HODECOD}{StandardiSed Disposition Term}
\item{HOPRESP}{Pre-Specified Healthcare Encounter}
\item{HOOCCUR}{Healthcare Encounter Occurrence}
\item{HOSTAT}{Completion Status}
\item{HOREASND}{Reason Healthcare Encounter Not Done}
\item{HODTC}{Date/Time of Specimen Collection}
\item{HOSTDTC}{Start Date/Time of Disposition Event}
\item{HOENDTC}{End Date/Time of Healthcare Encounter}
\item{HODY}{Study Day of Specimen Collection}
\item{HOSTDY}{Study Day of Start of Disposition Event}
\item{HOENDY}{Study Day of End of Healthcare Encounter}
\item{HODUR}{Duration of Healthcare Encounter}
\item{HOSTRF}{Start Relative to Reference Period}
\item{HOEVINTX}{Evaluation Interval Text}
\item{HOCDSTDY}{Contributor-Defined Study Day}
\item{HODISOUT}{Discharge Outcome}
\item{SELFCARE}{Ability to Self-Care at Discharge}
\item{HOINDC}{Indication}
\item{HOPATT}{Number of Healthcare Encounters}
}
}
}
\usage{
HO_RPTESTB
}
\description{
A fictional HO dataset for testing code and training
}
\keyword{datasets}
