% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_RPTESTB.R
\docType{data}
\name{PE_RPTESTB}
\alias{PE_RPTESTB}
\title{IDDO-SDTM Tutorial Physical Examination (PE) Domain}
\format{
\subsection{}{

A data frame with 12 rows and 22 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{PESEQ}{Sequence Number}
\item{PETESTCD}{Body System Examined Short Name}
\item{PETEST}{Body System Examined}
\item{PECAT}{Category for Examination}
\item{PESCAT}{Subcategory for Examination}
\item{PEBODSYS}{Body System or Organ Class}
\item{PEORRES}{Verbatim Examination Finding}
\item{PEORRESU}{Original Units}
\item{PESTRESC}{Character Result/ Finding in Std Format}
\item{PESTAT}{Completion Status}
\item{PEREASND}{Reason Not Examined}
\item{PELOC}{Location of Physical Exam Finding}
\item{PEMETHOD}{Method of Test or Examination}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EPOCH}{Epoch}
\item{PEDTC}{Date/Time of Examination}
\item{PEDY}{Study Day of Examination}
}
}
}
\usage{
PE_RPTESTB
}
\description{
A fictional PE dataset for testing code and training
}
\keyword{datasets}
