% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF_RPTESTB.R
\docType{data}
\name{PF_RPTESTB}
\alias{PF_RPTESTB}
\title{IDDO-SDTM Tutorial Pharmacogenomics Genetics (PF) Domain}
\format{
\subsection{}{

A data frame with 12 rows and 26 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{NHOID}{Non-Host Organism Identifier}
\item{PFSEQ}{Sequence Number}
\item{PFREFID}{Reference ID}
\item{PFTESTCD}{Short Name of Pharmacogenomics Lab Test}
\item{PFTEST}{Name of Pharmacogenomics Lab Test}
\item{PFGENRI}{Genetic Region of Interest}
\item{PFGENTRG}{Genetic Target}
\item{PFCAT}{Category for Pharmacogenomics Test}
\item{PFORRES}{Result or Finding in Original Units}
\item{PFORRESU}{Original Units}
\item{PFSTRESC}{Result or Finding in Standard Format}
\item{PFSTRESN}{Numeric Result/Finding in Standard Units}
\item{PFSTRESU}{Standard Units}
\item{PFSPEC}{Specimen Type}
\item{PFMETHOD}{Method Code for Test}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EPOCH}{Epoch}
\item{PFDTC}{Date/Time of Specimen Collection}
\item{PFDY}{Study Day of Specimen Collection}
\item{PFCDSTDY}{Contributor-Defined Study Day}
}
}
}
\usage{
PF_RPTESTB
}
\description{
A fictional PF dataset for testing code and training
}
\keyword{datasets}
