% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_domain.R
\name{prepare_domain}
\alias{prepare_domain}
\title{Prepare IDDO-SDTM domain for analysis}
\usage{
prepare_domain(
  data,
  domain,
  include_location = FALSE,
  include_method = FALSE,
  variables_include = c(),
  timing_variables = c(str_c(domain, "HR"), str_c(domain, "DY"), str_c(domain, "STDY"),
    str_c(domain, "CDSTDY"), "VISITDY", "EPOCH"),
  values_fn = first,
  print_messages = TRUE
)
}
\arguments{
\item{data}{Domain data frame.}

\item{domain}{Character. The two letter domain name of the data. Domain
options: "DM", "LB", "MB", "VS", "RS", "DD", "RP", "SC", "MP", "PF", "AU",
"PC", "SA", "HO", "ER", "PO", "DS}

\item{include_location}{Boolean. Should the finding/event location (--LOC) be
included in the output. Default is FALSE.}

\item{include_method}{Boolean. Should the finding/event method (--METHOD) be
included in the output. Default is FALSE.}

\item{variables_include}{Character list. List of variables to include in the
output. Default is to include all available variables.}

\item{timing_variables}{Character list. List of timing variables which are to
be used to separate time points, this is hierarchical so the order is taken
into account. Default is: --HR, --DY, --STDY, --CDSTDY, VISITDY, EPOCH.

(using default for example) Each row will be initially summarised based on
the --HR (study hour) variable, if that is missing then the --DY (study
day) variable is used, and so on. The output will be one row per
participant, per time point, where the time point for each row is the first
available variable listed in timing_variables.}

\item{values_fn}{Function. The function which will determine which data row
is used in the output, in the event there are multiple rows for the same
subject with the same time points (as listed in timing_variables). Default
is first(), i.e. if there is two rows from the same day and time, the first
record will be taken, the second will be dropped. Choice of
timing_variables will impact the number of rows affected.}

\item{print_messages}{Boolean. Should messages from the function be generated
and shown in the user's console. Default is TRUE.}
}
\value{
A dataframe which has been cleaned and subset based on the input
parameters.
}
\description{
Amalgamate domain data and pivot wider to convert the IDDO-SDTM format data
into a more analysable format. Function works on one domain and requires the
two letter domain name as well as the domain data file.
}
\examples{

prepare_domain(DM_RPTESTB, "DM")

# Select just ARMCD, AGE & SEX
prepare_domain(DM_RPTESTB, "DM", variables_include = c("ARMCD", "AGE", "SEX"))

# Change which timing_variables are used to summarise the data
prepare_domain(LB_RPTESTB, "lb", timing_variables = c("VISITNUM", "VISITDY"))

# Include location in the output and change the values_fn to select the last result
prepare_domain(VS_RPTESTB, "vs", include_location = TRUE, values_fn = dplyr::last)

}
