% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS_RPTESTB.R
\docType{data}
\name{RS_RPTESTB}
\alias{RS_RPTESTB}
\title{IDDO-SDTM Tutorial Disease Response and Clinical Classification (RS) Domain}
\format{
\subsection{}{

A data frame with 4 rows and 28 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{RSSEQ}{Sequence Number}
\item{RSTESTCD}{Assessment Short Name}
\item{RSTEST}{Assessment Name}
\item{RSCAT}{Category for Assessment}
\item{RSORRES}{Result or Finding in Original Units}
\item{RSORRESU}{Original Units}
\item{RSSTRESC}{Character Result/Finding in Std Format}
\item{RSSTRESN}{Numeric Result/Finding in Std Format}
\item{RSSTRESU}{Standard Units}
\item{RSSTAT}{Completion Status}
\item{RSREASND}{Reason Not Done}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EPOCH}{Epoch}
\item{RSDTC}{Date/Time of Assessment}
\item{RSDY}{Study Day of Assessment}
\item{RSTPT}{Planned Time Point Name}
\item{RSTPTNUM}{Planned Time Point Number}
\item{RSTPTREF}{Time Point Reference}
\item{RSSTRF}{Start Relative to Reference Period}
\item{RSEVINTX}{Evaluation Interval Text}
\item{RSSTRTPT}{Start Relative to Reference Time Point}
\item{RSSTTPT}{Start Reference Time Point}
\item{RSCDSTDY}{Contributor-Defined Study Day}
}
}
}
\usage{
RS_RPTESTB
}
\description{
A ficRSonal RS dataset for tesRSng code and training
}
\keyword{datasets}
