% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_variables.R
\name{table_variables}
\alias{table_variables}
\title{Tabulate function to display variables in a given domain}
\usage{
table_variables(domain, data, by_STUDYID = FALSE)
}
\arguments{
\item{domain}{Character. The two letter code for the domain which matches the data.
Character string. Domains included: "DM", "LB", "RP", "MB", "MP", "SA",
"IN", "VS", "DS", "RS", "PO", "SC", "HO", "ER" & "DD".}

\item{data}{Domain data frame.}

\item{by_STUDYID}{Boolean. Split data by STUDYID if TRUE. Default is FALSE.}
}
\value{
For Demographics (DM) domain, a character list with the column names.
For all other domains, a table class object listing the variables under
--TESTCD or INTRT, MPLOC, DSDECOD; where -- is the two letter domain name.
}
\description{
Uses table() to display the variables contained within an SDTM
formatted data frame. Additionally can be split by STUDYID to display the
options across multiple studies.
}
\examples{
table_variables("LB", LB_RPTESTB)

table_variables("VS", VS_RPTESTB, by_STUDYID = TRUE)

}
