% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TI_RPTESTB.R
\docType{data}
\name{TI_RPTESTB}
\alias{TI_RPTESTB}
\title{IDDO-SDTM Tutorial Trial Inclusion Exclusion Criteria (TI) Domain}
\format{
\subsection{}{

A data frame with 7 rows and 5 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{IETESTCD}{Incl/Excl Criterion Short Name}
\item{IETEST}{Inclusion/Exclusion Criterion}
\item{IECAT}{Inclusion/Exclusion Category}
}
}
}
\usage{
TI_RPTESTB
}
\description{
A fictional TI dataset for testing code and training
}
\keyword{datasets}
