% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TV_RPTESTB.R
\docType{data}
\name{TV_RPTESTB}
\alias{TV_RPTESTB}
\title{IDDO-SDTM Tutorial Trial Visits (TV) Domain}
\format{
\subsection{}{

A data frame with 20 rows and 8 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{ARMCD}{Planned Arm Code}
\item{ARM}{Description of Planned Arm}
\item{TVSTRL}{Visit Start Rule}
}
}
}
\usage{
TV_RPTESTB
}
\description{
A fictional TV dataset for testing code and training
}
\keyword{datasets}
