% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_age_to_years.R
\name{convert_age_to_years}
\alias{convert_age_to_years}
\title{Convert AGE to years.}
\usage{
convert_age_to_years(data, remove_AGEU = TRUE)
}
\arguments{
\item{data}{data frame containing the AGE and AGEU variables; typically the
Demographics (DM) domain.}

\item{remove_AGEU}{Boolean. Whether the AGEU variable should be removed from
the output. Default is TRUE.}
}
\value{
data frame with AGE in years as opposed to the original values.
}
\description{
Convert the AGE of all subjects to years and change the AGEU to "YEARS".
}
\examples{
DM_RPTESTB

convert_age_to_years(DM_RPTESTB)

}
