% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_age_to_years.R
\name{convert_age_to_years}
\alias{convert_age_to_years}
\title{Convert AGE to years.}
\usage{
convert_age_to_years(data)
}
\arguments{
\item{data}{data frame containing the AGE and AGEU variables; typically the
Demographics (DM) domain.}
}
\value{
data frame with AGE and AGEU in years as opposed to the original
values.
}
\description{
Convert the AGE of all subjects to years and change the AGEU to
"YEARS".
}
