% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clinical_hist_table.R
\name{create_clinical_hist_table}
\alias{create_clinical_hist_table}
\title{Creates table of medical history features}
\usage{
create_clinical_hist_table(
  dm_domain,
  sa_domain,
  timing_variables = c("SAHR", "SADY", "SASTDY", "VISITDY", "VISITNUM", "VISIT", "EPOCH",
    "SAEVLINT", "SAEVINTX"),
  values_funct = first
)
}
\arguments{
\item{dm_domain}{The name of the demographics (DM) domain in the global
environment.}

\item{sa_domain}{The name of the clinical and adverse events (SA) domain in
the global environment.}

\item{timing_variables}{List of timing variables which are to be used to
separate time points, this is hierarchical so the order is taken into
account. Default is: SAHR, SADY, SASTDY, VISITDY, VISITNUM, VISIT, EPOCH,
SAEVLINT, SAEVINTX.}

\item{values_funct}{The function which will determine which data row is used
in the output, in the event there are multiple rows for the same subject
with the same time points (as listed in timing_variables). Default is
first(), i.e. if there is two rows from the same day and time, the first
record will be taken, the second will be dropped. Choice of
timing_variables will impact the number of rows affected.}
}
\value{
An analysis dataset, one row per person, per timepoint.
}
\description{
Uses the IDDO-SDTM clinical and adverse events (SA) domain to create a single
dataset with various historic clinical information summarised. Historic
refers to events that occured before the study start.
}
