% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_malaria_pcr_table.R
\name{create_malaria_pcr_table}
\alias{create_malaria_pcr_table}
\title{Create table of malaria specific PCR data}
\usage{
create_malaria_pcr_table(pf_domain, rs_domain)
}
\arguments{
\item{pf_domain}{The name of the pharmacogenomics genetics (PF) domain in the
global environment.}

\item{rs_domain}{The name of the disease response and clinical classification
(RS) domain in the global environment.}
}
\value{
An analysis dataset, one row per person, per timepoint.
}
\description{
Joins the pharmacogenomics genetics (PF) and disease response and clinical
classification (RS)  IDDO-SDTM domains together to create a single data
}
