% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_participant_table.R
\name{create_participant_table}
\alias{create_participant_table}
\title{Create table of participant details and baseline characteristics.}
\usage{
create_participant_table(
  dm_domain,
  lb_domain = NULL,
  mb_domain = NULL,
  rp_domain = NULL,
  sc_domain = NULL,
  vs_domain = NULL
)
}
\arguments{
\item{dm_domain}{The name of the demographics (DM) domain in the global
environment.}

\item{lb_domain}{The name of the laboratory test results (LB) domain in the
global environment.}

\item{mb_domain}{The name of the microbiology (MB) domain in the global
environment.}

\item{rp_domain}{The name of the reproductive system findings (RP) domain in
the global environment.}

\item{sc_domain}{The name of the subject characteristics (SC) domain in the
global environment.}

\item{vs_domain}{The name of the vital signs (VS) domain in the global
environment.}
}
\value{
An analysis dataset, one row per participant.
}
\description{
Joins several IDDO-SDTM domains together to create a single dataset with
participant demographics, characteristics and baseline test results or
findings.
}
