% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_anthro_scores.R
\name{derive_anthro_scores}
\alias{derive_anthro_scores}
\title{Calculate WHO growth standards using the anthro() package.}
\usage{
derive_anthro_scores(data, age_in_years = TRUE)
}
\arguments{
\item{data}{data frame that contains the AGE_YEARS, WEIGHT & HEIGHT
variables, typically held in the Demographics (DM) and Vital Signs (VS)
domains. Domains should have prepare_domain applied beforehand.}

\item{age_in_years}{Boolean. Is the AGE column in years. Default is FALSE,
and function will call \code{convert_age_to_years} if FALSE, otherwise will not
convert age.}
}
\value{
data frame with only only 5 year olds included and additional columns
for WAZ, HAZ and WHZ scores, along with flags for each.
}
\description{
Calculates the Height for Age (HAZ), Weight for Age (WAZ) and Weight for
Height (WHZ) z-scores which are used to measure growth against the WHO
population standards. This is only for children less than 5 years old.
}
\details{
The function can be applied to a dataset with subjects greater than 5 years
old too, but those subjects will be filtered out.
}
\examples{
# Merge the DM (AGE_YEARS, SEX) and VS (WEIGHT, HEIGHT) domains together
data = merge(
  prepare_domain(DM_RPTESTB, "dm"),
  prepare_domain(VS_RPTESTB, "vs")
  )

derive_anthro_scores(data)

}
