% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddoverse-package.R
\docType{package}
\name{iddoverse-package}
\alias{iddoverse}
\alias{iddoverse-package}
\title{iddoverse: Converting IDDO-SDTM Data To Analysis Datasets}
\description{
Functions to convert IDDO-curated SDTM domains into analysis datasets, similar to the Analysis Data Model (ADaM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://infectious-diseases-data-observatory.github.io/iddoverse/}
  \item Report bugs at \url{https://github.com/Infectious-Diseases-Data-Observatory/iddoverse/issues}
}

}
\author{
\strong{Maintainer}: Rhys Peploe \email{rhyspeploe1998@gmail.com} (\href{https://orcid.org/0009-0001-1669-3716}{ORCID})

}
\keyword{internal}
