% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_domain.R
\name{prepare_domain}
\alias{prepare_domain}
\title{Prepare IDDO-SDTM domain for analysis}
\usage{
prepare_domain(
  domain,
  data,
  include_LOC = FALSE,
  include_METHOD = FALSE,
  variables_include = c(),
  timing_variables = c(str_c(domain, "HR"), str_c(domain, "DY"), str_c(domain, "STDY"),
    "VISITDY", "VISITNUM", "VISIT", "EPOCH", str_c(domain, "EVLINT"), str_c(domain,
    "EVINTX")),
  values_fn = first
)
}
\arguments{
\item{domain}{The two letter domain name of the data.}

\item{data}{The name of the data in the global environment.}

\item{include_LOC}{Boolean. Should the location (--LOC) be included in the
output. Default is FALSE.}

\item{include_METHOD}{Boolean. Should the method (--METHOD) be included in
the output. Default is FALSE.}

\item{variables_include}{List of variables to include in the output. Default
is to include all available variables.}

\item{timing_variables}{List of timing variables which are to be used to
separate time points, this is hierarchical so the order is taken into
account. Default is: --HR, --DY, --STDY, VISITDY, VISITNUM, VISIT, EPOCH,
--EVLINT, --EVINTX.}

\item{values_fn}{The function which will determine which data row is used in
the output, in the event there are multiple rows for the same subject with
the same time points (as listed in timing_variables). Default is first(),
i.e. if there is two rows from the same day and time, the first record will
be taken, the second will be dropped. Choice of timing_variables will
impact the number of rows affected.}
}
\value{
A dataframe which has been cleaned and subset based on the input
parameters
}
\description{
Amalgamate domain data and pivot wider to convert the IDDO-SDTM format data
into a more analysable format. Command works on one domain and requires the
two letter domain name as well as the domain data file.
}
